Add comments, move request classes to new file
