+++ Octokit/Clients/IRepositoryBranchesClient.cs
@@ -110,6 +110,16 @@ namespace Octokit
         Task<BranchProtectionSettings> GetBranchProtection(string owner, string name, string branch);
 
         /// <summary>
+        /// Get the branch protection settings for the specified branch />
+        /// </summary>
+        /// <remarks>
+        /// See the <a href="https://developer.github.com/v3/repos/branches/#get-branch-protection">API documentation</a> for more details
+        /// </remarks>
+        /// <param name="repositoryId">The Id of the repository</param>
+        /// <param name="branch">The name of the branch</param>
+        Task<BranchProtectionSettings> GetBranchProtection(int repositoryId, string branch);
+
+        /// <summary>
         /// Update the branch protection settings for the specified branch />
         /// </summary>
         /// <remarks>
@@ -122,6 +132,16 @@ namespace Octokit
         Task<BranchProtectionSettings> UpdateBranchProtection(string owner, string name, string branch, BranchProtectionSettingsUpdate update);
 
         /// <summary>
+        /// Update the branch protection settings for the specified branch />
+        /// </summary>
+        /// <remarks>
+        /// See the <a href="https://developer.github.com/v3/repos/branches/#update-branch-protection">API documentation</a> for more details
+        /// </remarks>
+        /// <param name="repositoryId">The Id of the repository</param>
+        /// <param name="update">Branch protection settings</param>
+        Task<BranchProtectionSettings> UpdateBranchProtection(int repositoryId, string branch, BranchProtectionSettingsUpdate update);
+
+        /// <summary>
         /// Remove the branch protection settings for the specified branch />
         /// </summary>
         /// <remarks>
@@ -131,5 +151,15 @@ namespace Octokit
         /// <param name="name">The name of the repository</param>
         /// <param name="branch">The name of the branch</param>
         Task<bool> DeleteBranchProtection(string owner, string name, string branch);
+
+        /// <summary>
+        /// Remove the branch protection settings for the specified branch />
+        /// </summary>
+        /// <remarks>
+        /// See the <a href="https://developer.github.com/v3/repos/branches/#remove-branch-protection">API documentation</a> for more details
+        /// </remarks>
+        /// <param name="repositoryId">The Id of the repository</param>
+        /// <param name="branch">The name of the branch</param>
+        Task<bool> DeleteBranchProtection(int repositoryId, string branch);
     }
 }
