+++ Octokit/Octokit-Mono.csproj
@@ -491,6 +491,7 @@
     <Compile Include="Models\Response\TimelineEventInfo.cs" />
     <Compile Include="Clients\IRepositoryBranchesClient.cs" />
     <Compile Include="Clients\RepositoryBranchesClient.cs" />
+    <Compile Include="Models\Request\BranchProtectionUpdate.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
 </Project>
