+++ Octokit/Octokit-MonoAndroid.csproj
@@ -502,6 +502,7 @@
     <Compile Include="Models\Response\TimelineEventInfo.cs" />
     <Compile Include="Clients\IRepositoryBranchesClient.cs" />
     <Compile Include="Clients\RepositoryBranchesClient.cs" />
+    <Compile Include="Models\Request\BranchProtectionUpdate.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Novell\Novell.MonoDroid.CSharp.targets" />
 </Project>
